coloquei o /search na roda index e corrigi um erro de escrita do firebase
